requirements.txt
absl-py==2.1.0
appnope==0.1.3
arviz==0.17.0
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
bcrypt==4.1.2
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorlover==0.3.0
comm==0.2.1
contourpy==1.2.0
cufflinks==0.17.3
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
dill==0.3.8
distlib==0.3.8
dnspython==2.6.1
executing==2.0.1
fastjsonschema==2.19.1
fastprogress==1.0.3
filelock==3.13.1
Flask==3.0.2
Flask-Login==0.6.3
flatbuffers==23.5.26
fonttools==4.47.2
fsspec==2023.12.2
gast==0.5.4
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.0
h5netcdf==1.3.0
h5py==3.10.0
idna==3.6
install==1.3.5
ipykernel==6.29.0
ipython==8.20.0
ipywidgets==8.1.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab-widgets==3.0.9
keras==2.15.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5.2
MarkupSafe==2.1.4
matplotlib==3.8.2
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
mpmath==1.3.0
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.0
parso==0.8.3
patsy==0.5.6
pexpect==4.9.0
pillow==10.2.0
pipenv==2023.12.1
platformdirs==4.1.0
plotly==5.18.0
prompt-toolkit==3.0.43
protobuf==4.23.4
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments==2.17.2
pymongo==4.6.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
pyzmq==25.1.2
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.17.1
rsa==4.9
scikit-learn==1.4.0
scipy==1.12.0
seaborn==0.13.2
semver==3.0.2
six==1.16.0
stack-data==0.6.3
statsmodels==0.14.1
sympy==1.12
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.15.0
termcolor==2.4.0
Theano-PyMC==1.1.2
threadpoolctl==3.2.0
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tornado==6.4
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
virtualenv==20.25.1
wcwidth==0.2.13
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wrapt==1.14.1
xarray==2024.1.1
xarray-einstats==0.7.0
